+++ test/system/underline_panel_test.rb
@@ -2,14 +2,20 @@
 
 require "application_system_test_case"
 
-class IntegrationUnderlineNavComponentTest < ApplicationSystemTestCase
-  def assert_underline_nav_rendered
+class IntegrationUnderlinePanelTest < ApplicationSystemTestCase
+  def assert_underline_panel_rendered
     assert_selector("tab-container") do
       assert_selector("div.UnderlineNav") do
-        assert_selector("div.UnderlineNav-body[role='tablist']") do
-          assert_selector("button.UnderlineNav-item[role='tab'][aria-selected='true']", text: "Tab 1")
-          assert_selector("button.UnderlineNav-item[role='tab']", text: "Tab 2")
-          assert_selector("button.UnderlineNav-item[role='tab']", text: "Tab 3")
+        assert_selector("ul.UnderlineNav-body[role='tablist']") do
+          assert_selector("li[role='presentation']") do
+            assert_selector("button.UnderlineNav-item[role='tab'][aria-selected='true']", text: "Tab 1")
+          end
+          assert_selector("li[role='presentation']") do
+            assert_selector("button.UnderlineNav-item[role='tab']", text: "Tab 2")
+          end
+          assert_selector("li[role='presentation']") do
+            assert_selector("button.UnderlineNav-item[role='tab']", text: "Tab 3")
+          end
         end
       end
       assert_selector("div[role='tabpanel']", text: "Panel 1")
@@ -37,13 +43,13 @@ class IntegrationUnderlineNavComponentTest < ApplicationSystemTestCase
   def test_renders_component
     with_preview(:default)
 
-    assert_underline_nav_rendered
+    assert_underline_panel_rendered
   end
 
   def test_changes_tabs_on_click
     with_preview(:default)
 
-    assert_underline_nav_rendered
+    assert_underline_panel_rendered
 
     assert_selects_tab(2)
     assert_selects_tab(3)