+++ CHANGELOG.md
@@ -212,6 +212,10 @@ The category for changes related to documentation, testing and tooling. Also, fo
 
 ### Breaking changes
 
+* Split `UnderlineNavComponent` into `UnderlineNav` and `UnderlinePanel`.
+
+    *Kate Higa*
+
 * Require tab with panels to have `panel_id` so `aria-controls` can be set.
 
     *Kate Higa*